Twitter feed as component worked
